@@ -1,6 +1,6 @@
 module github.com/markliederbach/qrkdns
 
-go 1.16
+go 1.25
 
 require (
 	github.com/cloudflare/cloudflare-go v0.25.0
